CHOI/feellog-project_03/pyproject.toml
[project]
name = "feellog-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch==2.7.1",
    "torchvision==0.22.1",
    "torchaudio==2.7.1",
    "jupyterlab",
    "ipywidgets==8.1.7",
    "jupyterlab-widgets==3.0.15",
    "widgetsnbextension==4.0.14",
    "matplotlib>=3.9.4",
    "numpy",
    "openai",
    "opencv-python-headless",
    "pandas>=2.3.1",
    "uvicorn[standard]",
    "albumentations>=2.0.8",
    "scikit-learn>=1.6.1",
    "mlflow",
    "optuna",
    "tqdm",
    "pillow",
    "pandas",
    "ffmpeg",
    "flask",
    "faster-whisper",
    "huggingface_hub",
    "transformers",
    "moviepy",
    "onnxruntime==1.19.2",
    "google-generativeai==0.8.5",
    "soundfile==0.13.1",
    "pyloudnorm>=0.1.1",
    "audiomentations>=0.42.0",
    "pyannote-audio>=3.3.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
